+++ test/producer_test.exs
@@ -1,10 +1,11 @@
 defmodule EctoJob.ProducerTest do
   use ExUnit.Case, async: true
+  alias Ecto.Adapters.SQL.Sandbox
   alias EctoJob.Producer
   alias EctoJob.Test.{JobQueue, Repo}
 
   setup do
-    :ok = Ecto.Adapters.SQL.Sandbox.checkout(Repo)
+    :ok = Sandbox.checkout(Repo)
 
     %{
       state: %Producer.State{
@@ -12,7 +13,7 @@ defmodule EctoJob.ProducerTest do
         schema: JobQueue,
         notifier: nil,
         demand: 0,
-        clock: fn -> DateTime.from_naive!(~N[2017-08-17T12:24:00Z], "Etc/UTC") end,
+        clock: fn -> DateTime.from_naive!(~N[2017-08-17T12:24:00.000000Z], "Etc/UTC") end,
         poll_interval: 60_000,
         reservation_timeout: 60_000,
         execution_timeout: 300_000,
@@ -35,7 +36,7 @@ defmodule EctoJob.ProducerTest do
     end
 
     test "When scheduled jobs can be activated", %{state: state} do
-      at = DateTime.from_naive!(~N[2017-08-17T12:23:34.0Z], "Etc/UTC")
+      at = DateTime.from_naive!(~N[2017-08-17T12:23:34.000000Z], "Etc/UTC")
       Repo.insert!(JobQueue.new(%{}, schedule: at))
 
       assert {:noreply, [%JobQueue{}], %{demand: 9}} =