+++ test/supervisor_test.exs
@@ -1,8 +1,9 @@
 defmodule EctoJob.SupervisorTest do
   use ExUnit.Case
+  alias EctoJob.Test.JobQueue
 
   test "start_link" do
-    {:ok, pid} = EctoJob.Test.JobQueue.start_link(repo: EctoJob.Test.Repo, max_demand: 25)
+    {:ok, pid} = JobQueue.start_link(repo: EctoJob.Test.Repo, max_demand: 25)
 
     assert [
              {EctoJob.WorkerSupervisor, _, :supervisor, [EctoJob.WorkerSupervisor]},
@@ -10,8 +11,8 @@ defmodule EctoJob.SupervisorTest do
              {Postgrex.Notifications, notifications_pid, :worker, [Postgrex.Notifications]}
            ] = Supervisor.which_children(pid)
 
-    assert Process.whereis(EctoJob.Test.JobQueue.Supervisor) == pid
-    assert Process.whereis(EctoJob.Test.JobQueue.Notifier) == notifications_pid
-    assert Process.whereis(EctoJob.Test.JobQueue.Producer) == producer_pid
+    assert Process.whereis(JobQueue.Supervisor) == pid
+    assert Process.whereis(JobQueue.Notifier) == notifications_pid
+    assert Process.whereis(JobQueue.Producer) == producer_pid
   end
 end